Auto-package html resources into virtual filesystem
At build time, walk the html resource folder and embed everything into
the binary. This allows us to adjust our serving logic to

* Check if it's a special path
* Try to serve from www-root
* 404 if not found in root

Since we are no longer checking every file individually, we now derive
the mimetype by the extension of the file. If anyone uses the wrong
extension, that's on them.

Approach taken is to generate a zig file with a flat list of paths that
were found at build time. These paths are symlinked into the cache and
embedded into the final source with @embedFile(). The files have to be
symlinked as embedFile cannot embed files that are in a different
directory. At runtime we can just walk the list of paths and try to
serve

Tried to follow zigisims, bundling hand written tools is referenced as
the preferred way of doing things at
https://ziglang.org/learn/build-system/#running-system-tools

AFAICT the caching of the build system is working correctly. From
observed behvaior...
* addDirectoryArg ensures that the build system checks that the list of
  files in the directory are the same, not that the contents of those
  files is the same
* We then have to write out a dependency file, that tells the build
  system which files from within that directory we actually cared to use