Parameterize clusterer initialization
K means clustering is heavily effected by the number of means chosen.
Provide this option in the web UI and propagate to the clusterer.

For now we just always present and send the num means option, but it is
only used in the k means case

Move clusterer instantiation out of the setClusterer() app method. This
way we can create a clusterer and lazily parse the params relevant to it
in the server. It is no longer valid to use clusterers after init
without a reset